findBugs

+++ concur/src-dbcache/dbcache/conf/impl/DbRuleServiceImpl.java
@@ -402,7 +402,7 @@ public class DbRuleServiceImpl implements DbRuleService {
 		String valueStr = new StringBuffer().append(valueOfServer)
 											.append(ID_MAX_VALUE_OF_AUTOINCR)
 											.toString();
-		return Long.valueOf(valueStr);
+		return Long.valueOf(valueStr).longValue();
 	}
 
 	/**
@@ -415,7 +415,7 @@ public class DbRuleServiceImpl implements DbRuleService {
 		String valueStr = new StringBuffer().append(valueOfServer)
 											.append(STR_VALUE_OF_AUTOINCR_ID_MIN_VALUE)
 											.toString();
-		return Long.valueOf(valueStr);
+		return Long.valueOf(valueStr).longValue();
 	}
 
 	/**
